Vercel dev support, switch back to yarn
