+++ .gitignore
@@ -23,3 +23,4 @@ pnpm-debug.log*
 *.sw?
 
 junit.xml
+.vercel
