+++ src/main.ts
@@ -1,0 +1,1 @@
+import '../example-src/main'
